btnaddminutes modification in both tabMinutesList and minutesedit
